+++ inst/tinytest/test_get_mi.R
@@ -27,12 +27,14 @@ fit_mi <- measurement_invariance_models(fit_config,
                                         progress = FALSE)
 expect_equal(length(fit_mi), 17)
 expect_true(all(c("config", "metric", "scalar") %in% names(fit_mi)))
+expect_true(setequal(c(57, 69, 81, 67, 79), unique(sapply(fit_mi, fitMeasures, "df"))))
 
 fit_mi2 <- measurement_invariance_models(fit_config,
                                          max_free = 0,
                                          progress = FALSE)
 expect_equal(length(fit_mi2), 3)
 expect_true(setequal(c("config", "metric", "scalar"), names(fit_mi2)))
+expect_true(setequal(c(57, 69, 81), unique(sapply(fit_mi2, fitMeasures, "df"))))
 
 fit_mi3 <- measurement_invariance_models(fit_config,
                                          scalar = FALSE,
@@ -40,6 +42,7 @@ fit_mi3 <- measurement_invariance_models(fit_config,
                                          progress = FALSE)
 expect_equal(length(fit_mi3), 23)
 expect_false("scalar" %in% names(fit_mi3))
+expect_true(setequal(c(57, 69, 67, 65), unique(sapply(fit_mi3, fitMeasures, "df"))))
 
 fit_mi4 <- measurement_invariance_models(fit_config,
                                          metric = FALSE,
@@ -47,11 +50,13 @@ fit_mi4 <- measurement_invariance_models(fit_config,
                                          progress = FALSE)
 expect_equal(length(fit_mi4), 10)
 expect_false("metric" %in% names(fit_mi4))
+expect_true(setequal(c(57, 81, 79), unique(sapply(fit_mi4, fitMeasures, "df"))))
 
 out <- model_set(sem_out = fit_mi4,
                  skip_check_sem_out = TRUE,
                  progress = FALSE,
                  parallel = FALSE)
-out
+expect_true(is.null(out$equivalent_clusters) ||
+            length(out$equivalent_clusters) == 0)
 }